refactor: Project structure and convention
