+++ Makefile
@@ -2,15 +2,19 @@
 query:
 	bazel query //...
 
+.PHONY: build
 build:
 	bazel build //...
 
+.PHONY: test
 test:
 	bazel test //...
 
+.PHONY: clean
 clean:
 	bazel clean --async
 
+.PHONY: update
 update: update_aspect_bazelrc update_python_requirements
 	echo "Running update..."
 
@@ -22,6 +26,29 @@ update_python_requirements:
 	bazel run //third_party/python:requirements_3_10.update
 	bazel run //third_party/python:requirements_3_9.update
 	bazel run //third_party/python:requirements_3_8.update
+
+
+
+test_libs:
+	bazel test //libs/...
+
+test_projects:
+	bazel test //projects/...
+
+test_py_calculator:
+	echo "TODO"
+
+test_flask_calculator:
+	echo "TODO"
+
+test_fastapi_echo:
+	bazel test //projects/py_echo_fastapi_app:webapp_test
+
+test_oci_py_helloworld:
+	bazel test //projects/py_helloworld_cli_app:test
+
+test_oci_py_helloworld_v2:
+	bazel test //projects/py_helloworld_v2_cli_app:test
 
 
 
@@ -44,26 +71,6 @@ run_oci_py_helloworld_v2:
 	bazel build //projects/py_helloworld_v2_cli_app:tarball
 	docker load --input `bazel cquery --output=files //projects/py_helloworld_v2_cli_app:tarball`
 	docker run --rm local/py_helloworld_v2_cli_app:latest
-
-
-
-test_libs:
-	bazel test //libs/...
-
-test_py_calculator:
-	echo "TODO"
-
-test_flask_calculator:
-	echo "TODO"
-
-test_fastapi_echo:
-	bazel test //projects/py_echo_fastapi_app:webapp_test
-
-test_oci_py_helloworld:
-	bazel test //projects/py_helloworld_cli_app:test
-
-test_oci_py_helloworld_v2:
-	bazel test //projects/py_helloworld_v2_cli_app:test
