+++ libs/calculator/README.md
@@ -1,0 +1,4 @@
+# Overview
+
+Simple calculator library example.
+
