+++ projects/py_calculator_flask_app/BUILD.bazel
@@ -6,8 +6,8 @@ load("@rules_python//python:defs.bzl", "py_library")
 # install_deps()
 
 py_library(
-    name = "py_calculator_flask_app_lib",
-    srcs = ["src/app.py"],
+    name = "app_lib",
+    srcs = ["app/app.py"],
     deps = [
         requirement("Flask"),
         "//libs/calculator:calculator_lib",
@@ -15,10 +15,10 @@ py_library(
 )
 
 py_binary(
-    name = "app",
-    srcs = ["src/app.py"],
+    name = "app_bin",
+    srcs = ["app/app.py"],
     main = "app.py",
-    deps = [":py_calculator_flask_app_lib"],
+    deps = [":app_lib"],
 )
 
 # py_image(
