+++ projects/py_calculator_flask_app/app/app.py
@@ -1,14 +1,17 @@
+from flask import Flask
 from random import randint
 
-from libs.calculator.calculator import Calculator
+from libs.calculator.models.calculator import Calculator
 
+app = Flask(__name__)
 my_calculator = Calculator()
 
+@app.route('/')
 def hello():
   num1 = randint(0, 100)
   num2 = randint(0, 100)
   message = "Did you know {} + {} = {}?".format(num1, num2, my_calculator.add(num1, num2))
-  print(message)
+  return message
 
 if __name__ == '__main__':
-  hello()
+  app.run(host='0.0.0.0')
