--- projects/py_calculator_flask_app/tests/app_lib_test.py
@@ -1,24 +1,0 @@
-# Copyright 2023 The Bazel Authors. All rights reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-import os
-import sys
-
-import libs.app_lib as app_lib
-
-sanitized_version_check = f"{sys.version_info.major}_{sys.version_info.minor}"
-
-if not app_lib.websockets_is_for_python_version(sanitized_version_check):
-    print("expected package for Python version is different than returned")
-    sys.exit(1)
