+++ projects/py_echo_fastapi_app/app/web_app.py
@@ -1,9 +1,10 @@
 import click
 import logging
+
 from fastapi import FastAPI
 from fastapi.responses import JSONResponse
 
-logging.basicConfig(level=logging.INFO, format='[%(levelname)s] %(message)s')
+logging.basicConfig(level=logging.INFO, format='%(levelname)s: %(message)s')
 
 app = FastAPI()
 
@@ -11,6 +12,11 @@ app = FastAPI()
 @app.on_event("startup")
 async def startup_event():
     logging.info(f"=== [Starting FastAPI app...] ===")
+
+@app.on_event("shutdown")
+async def shutdown_event():
+    logging.info(f"=== [Stopping FastAPI app...] ===")
+
 
 @app.get("/")
 async def root():
