+++ projects/py_echo_fastapi_app/bin/run_bin.py
@@ -3,7 +3,7 @@ import uvicorn
 
 if __name__ == "__main__":
     # freeze_support()
-    uvicorn.run("projects.py_echo_fastapi_app.src.webapp:app", host="0.0.0.0", port=5000, reload=True, access_log=True)
+    uvicorn.run("projects.py_echo_fastapi_app.app.web_app:app", host="0.0.0.0", port=5000, reload=True, access_log=True)
 
     # sys.argv.insert(1, "projects.py_echo_fastapi_app.main:app")
     # sys.exit(uvicorn.main())  # pylint: disable=no-value-for-parameter
