+++ projects/py_helloworld_cli_app/app/__main__.py
@@ -1,4 +1,4 @@
-from projects.py_helloworld_cli_app.app import Helloworld
+from projects.py_helloworld_cli_app.app.cli import Helloworld
 
 if __name__ == "__main__":
     app = Helloworld("John")
