+++ projects/py_helloworld_v2_cli_app/BUILD.bazel
@@ -7,8 +7,8 @@ load("@container_structure_test//:defs.bzl", "container_structure_test")
 py_library(
     name = "hello_world_lib",
     srcs = [
-        "src/__init__.py",
-        "src/app.py",
+        "app/__init__.py",
+        "app/cli.py",
     ],
     imports = [".."],
     visibility = ["//:__subpackages__"],
@@ -16,7 +16,7 @@ py_library(
 
 py_binary(
     name = "hello_world_bin",
-    srcs = ["src/__main__.py"],
+    srcs = ["app/__main__.py"],
     imports = [".."],
     main = "__main__.py",
     visibility = ["//:__subpackages__"],
@@ -78,7 +78,7 @@ oci_tarball(
 # - 
 container_structure_test(
     name = "test",
-    configs = ["tests/test.yaml"],
+    configs = ["tests/container_test.yaml"],
     image = ":py_helloworld_image",
     timeout = "short"
 )
